proof setup
